Now works with tinsel-0.6 on one box, all 6 boards.
More can be done to streamline the compilation, but it works for now.
Using some tinsel config stuff so it sticks to the correct numbers.
Placement and contraction can be improved using number of boards etc.
At the basic level it works. Will need some testing to discover real improvements